+++ src/domain/imap/imap_handler.rs
@@ -12,9 +12,10 @@ use crate::{
 pub fn notify<'a, ImapService: ImapServiceInterface<'a>>(
     keepalive: u64,
     config: &Config,
+    account: &Account,
     imap: &mut ImapService,
 ) -> Result<()> {
-    imap.notify(config, keepalive)
+    imap.notify(config, account, keepalive)
 }
 
 pub fn watch<'a, ImapService: ImapServiceInterface<'a>>(